Don't refer back to a list; just tell the user they must have election_id.
This is probably not the right long-term solution.  Ideally, I think it'd be
useful if elections could be marked private or public, and elections that are
public would be listed automatically, and private ones wouldn't appear.

It seems there are some remnants in the code that make it look like someone
was going to make a step 0, which would list the elections from the database,
but that was never done.

So, I've hard-coded what I wanted here, in hopes someone else comes along and
does it up right.