Assure $anon_token_id is integer.
The code that currently calls this only ever passes in a value that was
returned by elec_verify_voted_token() and as such no security issue
existed here.

However, assuming this function might be used for someone else in
future, we might as well case the value back to integer "just in case".

HT Tobias Mueller <muelli@cryptobitch.de> who pointed this out.